#167: Profile view: Add link to department
