+++ mobile/src/people/people-department.tsx
@@ -6,7 +6,16 @@ import { EmployeesList } from './employees-list';
 import { AppState } from '../reducers/app.reducer';
 import { EmployeeMap, EmployeesStore } from '../reducers/organization/employees.reducer';
 import { Employee } from '../reducers/organization/employee.model';
-import { openEmployeeDetailsAction } from '../employee-details/employee-details-dispatcher';
+import { openEmployeeDetailsAction, CurrentDepartmentNavigationParams } from '../employee-details/employee-details-dispatcher';
+import { NavigationRoute, NavigationScreenProp, NavigationLeafRoute } from 'react-navigation';
+
+interface NavigationTyped<P> extends NavigationScreenProp<NavigationRoute> {
+    getParam: <T extends keyof P>(param: T, fallback?: P[T]) => P[T];
+}
+
+interface NavigationProps {
+    navigation: NavigationTyped<CurrentDepartmentNavigationParams>;
+}
 
 interface PeopleDepartmentProps {
     employees: EmployeesStore;
@@ -14,14 +23,20 @@ interface PeopleDepartmentProps {
     employeesPredicate: (employee: Employee) => boolean;
 }
 
-const mapStateToProps = (state: AppState): PeopleDepartmentProps => {
+const mapStateToProps = (state: AppState, ownProps: NavigationProps): PeopleDepartmentProps => {
 
     const userEmployee = state.organization.employees.employeesById.get(state.userInfo.employeeId);
+    const departmentId = ownProps.navigation.getParam('departmentId', undefined);
 
     return {
         employees: state.organization.employees,
         userEmployee,
-        employeesPredicate: (employee: Employee) => userEmployee && employee.departmentId === userEmployee.departmentId
+        employeesPredicate: (employee: Employee) => {
+
+            return departmentId 
+                ? employee.departmentId === departmentId 
+                : userEmployee && employee.departmentId === userEmployee.departmentId;
+        }
     };
 };
 
@@ -32,8 +47,8 @@ const mapDispatchToProps = (dispatch: Dispatch<any>): EmployeesListDispatchProps
     onItemClicked: (employee: Employee) => dispatch(openEmployeeDetailsAction(employee))
 });
 
-export class PeopleDepartmentImpl extends React.Component<PeopleDepartmentProps & EmployeesListDispatchProps> {
-    public shouldComponentUpdate(nextProps: PeopleDepartmentProps & EmployeesListDispatchProps) {
+export class PeopleDepartmentImpl extends React.Component<PeopleDepartmentProps & EmployeesListDispatchProps & NavigationProps> {
+    public shouldComponentUpdate(nextProps: PeopleDepartmentProps & EmployeesListDispatchProps & NavigationProps) {
         if (this.props.onItemClicked !== nextProps.onItemClicked
             || this.props.userEmployee !== nextProps.userEmployee
         ) {
